refactor(runtime): events_schema を責務別 sub-module に分割（MEW-49）
ユーザー指摘: 1,359 行の単一ファイル (events_schema.rs) は全体把握が
厳しい。テストと実装が別れていなくてもいいが、実装が責務別に分かれ
ていると行数が減ってテスト↔実装の往復がしやすい。

events_schema.rs (1,359 行) を以下に分割:

- mod.rs (33 行): module 宣言 + pub use 再エクスポート + #[cfg(test)]
  mod tests
- types.rs (284 行): EventsSchema / EventDef / FieldSpec / Tier /
  FieldType / RangeBound と serde 周辺ヘルパ (deserialize_unique_*)
- loader.rs (88 行): load_from_path / resolve_events_yaml_path /
  LoadOutcome / LoadError
- validator.rs (331 行): validate / validate_event / validate_field_spec /
  validate_range / validate_default / yaml_to_f64 / ValidationError
- tests.rs (623 行): 39 件の unit test 全て (#[cfg(test)] mod tests;
  経由)

可視性:
- types の構造体・enum は pub
- FieldType の internal method (supports_range / supports_max_length /
  is_filterable / default_range) は pub(super) で sub-module に公開
- FieldType::parse / is_scalar は types.rs 内のみで使用するため private
- deserialize_unique_* は types.rs 内のみで使用するため private

#![allow(dead_code, unused_imports)] を mod.rs 1 か所に集約（pub use 再
エクスポートが main.rs から使われていないため unused_imports も抑制）。

unit test 39 件全 pass。clippy / fmt clean。
